move helper function to helpers
