+++ app-web/src/components/GithubTemplate/Actions/Actions.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled from '@emotion/styled';
-import { getGithubIssuesRoute } from '../../../utils/helpers';
-import { Link } from '../../UI/Link';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faGithub } from '@fortawesome/fontawesome-free-brands';
+import { getCannedIssueMessage } from '../../../utils/helpers';
+import { Link } from '../../UI/Link';
 
 const Container = styled.ul`
   display: flex;
@@ -26,24 +26,6 @@ const LI = styled.li`
     text-decoration: underline;
   }
 `;
-
-/**
- * gets a github issue route with a precanned message as params
- * @param {String} repo
- * @param {String} owner
- * @param {String} pageTitle the title of the resource page
- * @param {String} originalSource path to the markdown file as found in github
- * @param {String} devhubPath path to the resource page as found in devhub
- * @returns {String} the url to the issues route
- */
-export const getCannedIssueMessage = (repo, owner, pageTitle, originalSource, devhubPath) => {
-  const route = getGithubIssuesRoute(repo, owner);
-  const title = encodeURIComponent(`Devhub Issue: ${pageTitle} [short description here]`);
-  const body = encodeURIComponent(
-    `> path: (do not delete) ${originalSource}\n > (do not delete) devhub page: ${devhubPath}\n\n## Devhub Content Issue\n[description of your issue here]`,
-  );
-  return `${route}/new?title=${title}&body=${body}`;
-};
 
 export const IDS = {
   issue: 'actions-issue',
